Add files via upload
Introduction:
A CLI Based project with the main functionalities of ATM machine. It also Includes the administration control part to show some featues like an actual ATM Machine. I is developed in C++. 

Objective:
The main objective of the ATM machine project is to provide a way to made this system using OOP concepts. This will provide the concepts of the OOP here. Moreover the were to provide the functionalities of withdraw, deposit, change pin, display complete data, display specific account information.

Features and Explaination:
This project is about the ATM machine. The main functionalities of ATM will be described here. This project is a based on the OOP concepts. Its lines are 624. Security is the main consideration in this project. Admin pin is for the admin(head) control. This will provide 7 options: Account details, Details of account, Check details of account, Withdraw, Change pin, Exit.

Future Extension:
We can Extend this project to GUI based project with advanced functionality. And we can use this concept for some modern frameworks.